tests/wpt/web-platform-tests/css/css-writing-modes/parsing/text-combine-upright-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Writing Modes Level 3: parsing text-combine-upright with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-writing-modes-3/#propdef-text-combine-upright">
<meta name="assert" content="text-combine-upright supports only the grammar 'none | all'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("text-combine-upright", "auto");
test_invalid_value("text-combine-upright", "none all");
</script>
</body>
</html>
